Fix: Ensure particles render by moving background init logic
I moved the background initialization logic, including the mousemove event listener for the glow effect, from App.vue into FancyBackground.vue.

This change ensures that the DOM elements within FancyBackground.vue are fully rendered before any attempts to manipulate them, which was the likely cause of the particles not appearing.

The `initializeBackground` function is now called within the `onMounted` lifecycle hook of `FancyBackground.vue`, and event listeners are cleaned up in `onBeforeUnmount`.
